+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- InputDeviceType enum
 - Device Input Visualizer component
 - Device DisplaySettings for PS4, XBOX360 and Keyboard&Mouse Dark/Light
 - JSON input config for Keyboard&Mouse