+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - InputDeviceType enum
 - Device Input Visualizer component
-- Device DisplaySettings for PS4, XBOX360 and Keyboard&Mouse Dark/Light
+- Device Display Settings for PS3, PS4, Xbox360, Xbox360, XboxSeriesX, Keyboard&Mouse Dark/Light, WebGLGamepad and Generic Gamepads
 - JSON input config for Keyboard&Mouse
 - JSON input config for DualMotorGamepad
 - Sprites from Xelu's FreePromptsPack