+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.0-preview.2] - 2021-01-04
 ### Added
 - InputDeviceType enum
 - Device Input Visualizer component
@@ -27,5 +29,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - README
 - Initial commit
 
-[Unreleased]: https://bitbucket.org/nostgameteam/input-system/branches/compare/master%0D1.0.0-preview.1
+[Unreleased]: https://bitbucket.org/nostgameteam/input-system/branches/compare/master%0D1.0.0-preview.2
+[1.0.0-preview.2]: https://bitbucket.org/nostgameteam/input-system/src/1.0.0-preview.2/
 [1.0.0-preview.1]: https://bitbucket.org/nostgameteam/input-system/src/1.0.0-preview.1/